Consolidate utility
