+++ src/RogueCollectionQml/RoguePlugin/utility_qml.h
@@ -1,0 +1,6 @@
+#pragma once
+#include <string>
+
+std::string GetResourcePath(const std::string& filename);
+void DisplayMessage(const std::string& type, const std::string& title, const std::string& msg);
+void QuitApplication();
