+++ src/Shared/Frontend/utility.h
@@ -26,7 +26,8 @@ std::string GetTimestamp();
 uint32_t CharText(uint32_t ch);
 uint32_t CharColor(uint32_t ch);
 bool IsText(uint32_t ch);
-char GetUnixChar(unsigned char c);
-unsigned int GetColor(int chr, int attr);
-unsigned char FlipColor(unsigned char c);
+bool BlinkChar(uint32_t ch);
+uint32_t GetUnixChar(uint32_t c);
+uint32_t GetTileColor(int chr, int attr);
+uint32_t FlipColor(uint32_t c);
 char GetRawCharFromData(uint32_t* data, int r, int c, int cols);
